Move data.{c,h} under client/
